chore: Update choose_location.dart
Refactor the code in choose_location.dart to improve readability and maintainability. This commit reflects the necessary changes to ensure the code follows best practices for code organization and readability.
