+++ lib/modules/task.dart
@@ -1,0 +1,7 @@
+class Task {
+  final String title;
+  final String logo;
+  final String url;
+
+  Task({required this.title, required this.logo, required this.url});
+}
